+++ framework/domain/src/main/java/is/codion/framework/domain/entity/DefaultEntityDefinition.java
@@ -27,6 +27,7 @@ import is.codion.framework.domain.entity.attribute.DerivedAttributeDefinition;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
 import is.codion.framework.domain.entity.attribute.ForeignKey.Reference;
 import is.codion.framework.domain.entity.attribute.ForeignKeyDefinition;
+import is.codion.framework.domain.entity.attribute.ValueAttributeDefinition;
 import is.codion.framework.domain.entity.condition.ConditionString;
 import is.codion.framework.domain.entity.condition.ConditionType;
 import is.codion.framework.domain.entity.query.EntitySelectQuery;
@@ -764,7 +765,7 @@ final class DefaultEntityDefinition implements EntityDefinition, Serializable {
 																							Map<ForeignKey, List<ColumnDefinition<?>>> foreignKeyColumnDefinitions) {
 			//make foreign keys nullable if and only if any of their constituent columns are nullable
 			foreignKeyBuilder.nullable(foreignKeyColumnDefinitions.get(foreignKeyBuilder.attribute()).stream()
-							.anyMatch(AttributeDefinition::nullable));
+							.anyMatch(ValueAttributeDefinition::nullable));
 		}
 	}